requirements.txt
qdrant-client>=1.9.0
sentence-transformers>=2.7.0
openai>=1.30.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
PyPDF2>=3.0.0
pytest>=8.0.0
pytest-mock>=3.14.0
